+++ cocoa-beans-commands/src/main/java/net/apartium/cocoabeans/commands/SimpleArgumentMapper.java
@@ -105,14 +105,14 @@ public class SimpleArgumentMapper implements ArgumentMapper {
     }
 
     @Override
-    public List<ArgumentIndex<?>> mapIndices(RegisteredVariant.Parameter[] parameters, List<RegisterArgumentParser<?>> argumentParsers, List<Requirement> requirements, List<Class<?>> additionalTypes) {
+    public List<ArgumentIndex<?>> mapIndices(RegisteredVariant.Parameter[] parameters, List<RegisterArgumentParser<?>> argumentParsers, List<Requirement> requirements, Map<Class<?>,  List<ArgumentIndex<?>>> providedArgumentIndexesByType) {
         if (parameters.length == 0)
             return List.of();
 
         List<ArgumentIndex<?>> result = new ArrayList<>(parameters.length);
 
         Map<Class<?>, Integer> counterMap = new HashMap<>();
-        ResultMap resultMap = createParsedArgs(argumentParsers, requirements, additionalTypes, getParametersNames(parameters));
+        ResultMap resultMap = createParsedArgs(argumentParsers, requirements, getParametersNames(parameters), providedArgumentIndexesByType);
 
         for (RegisteredVariant.Parameter parameter : parameters) {
             Class<?> type = parameter.type();
@@ -337,7 +337,7 @@ public class SimpleArgumentMapper implements ArgumentMapper {
         return mapOfArguments.get(Sender.class).get(index - 1);
     }
 
-    private ResultMap createParsedArgs(List<RegisterArgumentParser<?>> argumentParsers, List<Requirement> requirements, List<Class<?>> additionalTypes, Map<String, Class<?>> lookupParametersNames) {
+    private ResultMap createParsedArgs(List<RegisterArgumentParser<?>> argumentParsers, List<Requirement> requirements, Map<String, Class<?>> lookupParametersNames, Map<Class<?>, List<ArgumentIndex<?>>> providedArgumentIndexesByType) {
         Map<Class<?>, List<ArgumentIndex<?>>> resultMap = new HashMap<>();
         Map<String, ArgumentIndex<?>> resultParameterName = new HashMap<>();
 
@@ -355,9 +355,14 @@ public class SimpleArgumentMapper implements ArgumentMapper {
             }
         }
 
-        for (Class<?> type : additionalTypes) {
-            serializesArgumentIndex(type, null, lookupParametersNames, counterMap, resultMap, resultParameterName);
-        }
+        providedArgumentIndexesByType.forEach((type, indices) -> indices
+                .forEach(index -> {
+                    int countIndex = counterMap.getOrDefault(type, 0);
+                    counterMap.put(type, countIndex + 1);
+
+                    resultMap.computeIfAbsent(type, k -> new ArrayList<>()).add(index);
+                })
+        );
 
         return new ResultMap(resultMap, resultParameterName);
     }
@@ -397,20 +402,6 @@ public class SimpleArgumentMapper implements ArgumentMapper {
             }
 
             resultParameterName.put(parameterName, context -> context.parsedArgs().get(type).get(countIndex));
-            return;
-        }
-
-        if (CommandException.class.isAssignableFrom(type)) {
-            resultMap.computeIfAbsent(type, k -> new ArrayList<>())
-                    .add(context -> {
-                        for (Map.Entry<Class<?>, List<Object>> entry : context.parsedArgs().entrySet()) {
-                            if (type.isAssignableFrom(entry.getKey())) {
-                                return entry.getValue().get(0);
-                            }
-                        }
-
-                        return null;
-                    });
             return;
         }